feat: refactor to remove duplicity
